# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenrouterConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenrouterMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenrouterConversationBody {
  title: string;
}

export interface SendOpenrouterMessageBody {
  content: string;
  model?: string;
  /** @nullable */
  maxTokens?: number | null;
  /** @nullable */
  temperature?: number | null;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  apiUrl?: string | null;
  skipAiCompletion?: boolean;
}

export interface OpenrouterConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenrouterMessage[];
}

export interface UpdateOpenrouterMessageBody {
  content: string;
}

export interface OpenrouterError {
  error: string;
}