# lib/api-zod/src/generated/types/sendOpenrouterMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendOpenrouterMessageBody {
  content: string;
  model?: string;
  /** @nullable */
  maxTokens?: number | null;
  /** @nullable */
  temperature?: number | null;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  apiUrl?: string | null;
  skipAiCompletion?: boolean;
}